requirements.txt
# Dependencias principales
openai>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
moviepy>=1.0.3
pillow>=10.0.0
edge-tts>=6.1.0
youtube-dl>=2021.12.17
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0

# Base de datos
sqlite3
sqlalchemy>=2.0.0

# Scraping y análisis
beautifulsoup4>=4.12.0
selenium>=4.15.0
pytrends>=4.9.0

# Procesamiento de imágenes y video
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0

# APIs de terceros
unsplash-python>=1.0.0
elevenlabs>=0.2.0

# Utilidades
schedule>=1.2.0
python-crontab>=2.7.0
rich>=13.0.0
typer>=0.9.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
mock>=5.1.0

# N8N alternativo (opcional)
# n8n-python-client